+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 - fix(nextjs): Don't add '.env.sentry-build-plugin' to .gitignore if it's already there (#610)
+- feat(nextjs): Support all `next.config` file types (#630)
+- fix(nextjs): Update instrumentation and example creation logic for app or pages usage (#629)
+- feat(nextjs): Prompt for `reactComponentAnnotation` (#634)
+- fix(nextjs): Add missing Error.getInitialProps calls in Next.js error page snippets (#632)
+- fix/feat: Improve error logging for package installation (#635)
 
 ## 3.25.2